refactor(SharingDetailsModal): Group with his parent ShareModal
