+++ packages/cozy-sharing/src/components/ShareModal/SharingDetailsModal.jsx
@@ -5,10 +5,10 @@ import { Dialog } from 'cozy-ui/transpiled/react/CozyDialogs'
 import Divider from 'cozy-ui/transpiled/react/Divider'
 import { useI18n } from 'cozy-ui/transpiled/react/providers/I18n'
 
-import OwnerIdentity from './components/Recipient/OwnerIdentity'
-import WhoHasAccess from './components/WhoHasAccess'
-import { getDisplayName } from './models'
 import styles from './share.styl'
+import { getDisplayName } from '../../models'
+import OwnerIdentity from '../Recipient/OwnerIdentity'
+import WhoHasAccess from '../WhoHasAccess'
 
 export const SharingDetailsModal = ({
   onClose,
